+++ app/src/main/java/ch/epfl/sweng/zuluzulu/User/AuthenticatedUser.java
@@ -24,12 +24,12 @@ public class AuthenticatedUser extends User {
     private final String last_names;
     // All followed ids of Associations, Chats and Events
     private List<Long> fav_assos;
-    private List<Integer> followed_chats;
-    private List<Integer> followed_events;
+    private List<Long> followed_chats;
+    private List<Long> followed_events;
     // TODO add argument to constructor and store liked_event in firebase
-    private Set<Integer> liked_events = new HashSet<>();
+    private Set<Long> liked_events = new HashSet<>();
 
-    protected AuthenticatedUser(String sciper, String gaspar, String email, String section, String semester, String first_names, String last_names, List<Long> fav_assos, List<Integer> followed_events, List<Integer> followed_chats) {
+    protected AuthenticatedUser(String sciper, String gaspar, String email, String section, String semester, String first_names, String last_names, List<Long> fav_assos, List<Long> followed_events, List<Long> followed_chats) {
         firestore_path = "users_info/" + sciper;
         this.sciper = sciper;
         this.gaspar = gaspar;
@@ -73,7 +73,7 @@ public class AuthenticatedUser extends User {
         Utils.removeIdFromList(firestore_path, "fav_assos", asso.getId());
     }
 
-    public void setFollowedChats(List<Integer> followed_chats) {
+    public void setFollowedChats(List<Long> followed_chats) {
         this.followed_chats = followed_chats;
     }
 
@@ -81,23 +81,23 @@ public class AuthenticatedUser extends User {
         this.fav_assos = fav_assos;
     }
 
-    public void setFollowedEvents(List<Integer> followed_events) {
+    public void setFollowedEvents(List<Long> followed_events) {
         this.followed_events = followed_events;
     }
 
-    public void setLikedEvent(Set<Integer> liked_events) {
+    public void setLikedEvent(Set<Long> liked_events) {
         this.liked_events = liked_events;
     }
 
-    public boolean isEventLiked(Integer event) {
+    public boolean isEventLiked(Long event) {
         return liked_events.contains(event);
     }
 
-    public void likeEvent(Integer event) {
+    public void likeEvent(Long event) {
         this.liked_events.add(event);
     }
 
-    public void dislikeEvent(Integer event) {
+    public void dislikeEvent(Long event) {
         this.liked_events.remove(event);
     }
